+++ group_test.go
@@ -18,28 +18,28 @@ func TestGetGroups(t *testing.T) {
 		t.Logf("  Name: %s", g.Name)
 		t.Logf("  Lights: %s", g.Lights)
 		t.Logf("  Type: %s", g.Type)
-		t.Logf("  State")
-		t.Logf("    AllOn: %t", g.State.AllOn)
-		t.Logf("    AnyOn: %t", g.State.AnyOn)
+		t.Logf("  GroupState:")
+		t.Logf("    AllOn: %t", g.GroupState.AllOn)
+		t.Logf("    AnyOn: %t", g.GroupState.AnyOn)
 		t.Logf("  Recycle: %t", g.Recycle)
 		t.Logf("  Class: %s", g.Class)
-		t.Logf("  Action:")
-		t.Logf("    On: %t", g.Action.On)
-		t.Logf("    Bri: %d", g.Action.Bri)
-		t.Logf("    Hue: %d", g.Action.Hue)
-		t.Logf("    Sat: %d", g.Action.Sat)
-		t.Logf("    Xy: %b", g.Action.Xy)
-		t.Logf("    Ct: %d", g.Action.Ct)
-		t.Logf("    Alert: %s", g.Action.Alert)
-		t.Logf("    Effect: %s", g.Action.Effect)
-		t.Logf("    TransitionTime: %d", g.Action.TransitionTime)
-		t.Logf("    BriInc: %d", g.Action.BriInc)
-		t.Logf("    SatInc: %d", g.Action.SatInc)
-		t.Logf("    HueInc: %d", g.Action.HueInc)
-		t.Logf("    CtInc: %d", g.Action.CtInc)
-		t.Logf("    XyInc: %d", g.Action.XyInc)
-		t.Logf("    ColorMode: %s", g.Action.ColorMode)
-		t.Logf("    Reachable: %t", g.Action.Reachable)
+		t.Logf("  State:")
+		t.Logf("    On: %t", g.State.On)
+		t.Logf("    Bri: %d", g.State.Bri)
+		t.Logf("    Hue: %d", g.State.Hue)
+		t.Logf("    Sat: %d", g.State.Sat)
+		t.Logf("    Xy: %b", g.State.Xy)
+		t.Logf("    Ct: %d", g.State.Ct)
+		t.Logf("    Alert: %s", g.State.Alert)
+		t.Logf("    Effect: %s", g.State.Effect)
+		t.Logf("    TransitionTime: %d", g.State.TransitionTime)
+		t.Logf("    BriInc: %d", g.State.BriInc)
+		t.Logf("    SatInc: %d", g.State.SatInc)
+		t.Logf("    HueInc: %d", g.State.HueInc)
+		t.Logf("    CtInc: %d", g.State.CtInc)
+		t.Logf("    XyInc: %d", g.State.XyInc)
+		t.Logf("    ColorMode: %s", g.State.ColorMode)
+		t.Logf("    Reachable: %t", g.State.Reachable)
 		t.Logf("  Id: %d", g.Id)
 	}
 }
@@ -58,28 +58,28 @@ func TestGetGroup(t *testing.T) {
 		t.Logf("Name: %s", g.Name)
 		t.Logf("Lights: %s", g.Lights)
 		t.Logf("Type: %s", g.Type)
-		t.Logf("State")
-		t.Logf("  AllOn: %t", g.State.AllOn)
-		t.Logf("  AnyOn: %t", g.State.AnyOn)
+		t.Logf("GroupState:")
+		t.Logf("  AllOn: %t", g.GroupState.AllOn)
+		t.Logf("  AnyOn: %t", g.GroupState.AnyOn)
 		t.Logf("Recycle: %t", g.Recycle)
 		t.Logf("Class: %s", g.Class)
-		t.Logf("Action:")
-		t.Logf("  On: %t", g.Action.On)
-		t.Logf("  Bri: %d", g.Action.Bri)
-		t.Logf("  Hue: %d", g.Action.Hue)
-		t.Logf("  Sat: %d", g.Action.Sat)
-		t.Logf("  Xy: %b", g.Action.Xy)
-		t.Logf("  Ct: %d", g.Action.Ct)
-		t.Logf("  Alert: %s", g.Action.Alert)
-		t.Logf("  Effect: %s", g.Action.Effect)
-		t.Logf("  TransitionTime: %d", g.Action.TransitionTime)
-		t.Logf("  BriInc: %d", g.Action.BriInc)
-		t.Logf("  SatInc: %d", g.Action.SatInc)
-		t.Logf("  HueInc: %d", g.Action.HueInc)
-		t.Logf("  CtInc: %d", g.Action.CtInc)
-		t.Logf("  XyInc: %d", g.Action.XyInc)
-		t.Logf("  ColorMode: %s", g.Action.ColorMode)
-		t.Logf("  Reachable: %t", g.Action.Reachable)
+		t.Logf("State:")
+		t.Logf("  On: %t", g.State.On)
+		t.Logf("  Bri: %d", g.State.Bri)
+		t.Logf("  Hue: %d", g.State.Hue)
+		t.Logf("  Sat: %d", g.State.Sat)
+		t.Logf("  Xy: %b", g.State.Xy)
+		t.Logf("  Ct: %d", g.State.Ct)
+		t.Logf("  Alert: %s", g.State.Alert)
+		t.Logf("  Effect: %s", g.State.Effect)
+		t.Logf("  TransitionTime: %d", g.State.TransitionTime)
+		t.Logf("  BriInc: %d", g.State.BriInc)
+		t.Logf("  SatInc: %d", g.State.SatInc)
+		t.Logf("  HueInc: %d", g.State.HueInc)
+		t.Logf("  CtInc: %d", g.State.CtInc)
+		t.Logf("  XyInc: %d", g.State.XyInc)
+		t.Logf("  ColorMode: %s", g.State.ColorMode)
+		t.Logf("  Reachable: %t", g.State.Reachable)
 		t.Logf("Id: %d", g.Id)
 		break
 	}
@@ -106,7 +106,7 @@ func TestCreateGroup(t *testing.T) {
 func TestUpdateGroup(t *testing.T) {
 	hue := huego.New(os.Getenv("HUE_HOSTNAME"), os.Getenv("HUE_USERNAME"))
 	id := 3
-	resp, err := hue.UpdateGroup(id, &huego.Group{
+	resp, err := hue.UpdateGroup(id, huego.Group{
 		Name: "TestGroup (Updated)",
 		Class: "Office",
 	})
@@ -123,7 +123,7 @@ func TestUpdateGroup(t *testing.T) {
 func TestSetGroupState(t *testing.T) {
 	hue := huego.New(os.Getenv("HUE_HOSTNAME"), os.Getenv("HUE_USERNAME"))
 	id := 3
-	resp, err := hue.SetGroupState(id, &huego.State{
+	resp, err := hue.SetGroupState(id, huego.State{
 		On: true,
 		Bri: 150,
 		Sat: 210,
@@ -147,4 +147,184 @@ func TestDeleteGroup(t *testing.T) {
 	} else {
 		t.Logf("Deleted group with id: %d", id)
 	}
+}
+
+func TestRenameGroup(t *testing.T) {
+	bridge := huego.New(os.Getenv("HUE_HOSTNAME"), os.Getenv("HUE_USERNAME"))
+	id := 3
+	group, err := bridge.GetGroup(id)
+	if err != nil {
+		t.Fatal(err)
+	}
+	err = group.Rename("MyGroup (renamed)")
+	if err != nil {
+		t.Fatal(err)
+	}
+	t.Logf("Group renamed to %s", group.Name)
+}
+
+func TestTurnOffGroup(t *testing.T) {
+	b := huego.New(os.Getenv("HUE_HOSTNAME"), os.Getenv("HUE_USERNAME"))
+	id := 1
+	group, err := b.GetGroup(id)
+	if err != nil {
+		t.Fatal(err)
+	}
+	err = group.Off()
+	if err != nil {
+		t.Fatal(err)
+	}
+	t.Logf("Turned off group with id %d", group.Id)
+	t.Logf("Group IsOn: %t", group.State.On)
+}
+
+func TestTurnOnGroup(t *testing.T) {
+	b := huego.New(os.Getenv("HUE_HOSTNAME"), os.Getenv("HUE_USERNAME"))
+	id := 1
+	group, err := b.GetGroup(id)
+	if err != nil {
+		t.Fatal(err)
+	}
+	err = group.On()
+	if err != nil {
+		t.Fatal(err)
+	}
+	t.Logf("Turned on group with id %d", group.Id)
+	t.Logf("Group IsOn: %t", group.State.On)
+}
+
+func TestIfGroupIsOn(t *testing.T) {
+	b := huego.New(os.Getenv("HUE_HOSTNAME"), os.Getenv("HUE_USERNAME"))
+	id := 1
+	group, err := b.GetGroup(id)
+	if err != nil {
+		t.Fatal(err)
+	}
+	t.Logf("Is group %d on?: %t", group.Id, group.IsOn())
+}
+
+func TestSetGroupBri(t *testing.T) {
+	b := huego.New(os.Getenv("HUE_HOSTNAME"), os.Getenv("HUE_USERNAME"))
+	id := 1
+	group, err := b.GetGroup(id)
+	if err != nil {
+		t.Fatal(err)
+	}
+	err = group.Bri(254)
+	if err != nil {
+		t.Fatal(err)
+	}
+	t.Logf("Brightness of group %d set to %d", group.Id, group.State.Bri)
+}
+
+func TestSetGroupHue(t *testing.T) {
+	b := huego.New(os.Getenv("HUE_HOSTNAME"), os.Getenv("HUE_USERNAME"))
+	id := 1
+	group, err := b.GetGroup(id)
+	if err != nil {
+		t.Fatal(err)
+	}
+	err = group.Hue(65535)
+	if err != nil {
+		t.Fatal(err)
+	}
+	t.Logf("Hue of group %d set to %d", group.Id, group.State.Hue)
+}
+
+func TestSetGroupSat(t *testing.T) {
+	b := huego.New(os.Getenv("HUE_HOSTNAME"), os.Getenv("HUE_USERNAME"))
+	id := 1
+	group, err := b.GetGroup(id)
+	if err != nil {
+		t.Fatal(err)
+	}
+	err = group.Sat(254)
+	if err != nil {
+		t.Fatal(err)
+	}
+	t.Logf("Sat of group %d set to %d", group.Id, group.State.Sat)
+}
+
+func TestSetGroupXy(t *testing.T) {
+	b := huego.New(os.Getenv("HUE_HOSTNAME"), os.Getenv("HUE_USERNAME"))
+	id := 1
+	group, err := b.GetGroup(id)
+	if err != nil {
+		t.Fatal(err)
+	}
+	err = group.Xy([]float32{0.1, 0.5})
+	if err != nil {
+		t.Fatal(err)
+	}
+	t.Logf("Xy of group %d set to %d", group.Id, group.State.Xy)
+}
+
+func TestSetGroupCt(t *testing.T) {
+	b := huego.New(os.Getenv("HUE_HOSTNAME"), os.Getenv("HUE_USERNAME"))
+	id := 1
+	group, err := b.GetGroup(id)
+	if err != nil {
+		t.Fatal(err)
+	}
+	err = group.Ct(16)
+	if err != nil {
+		t.Fatal(err)
+	}
+	t.Logf("Ct of group %d set to %d", group.Id, group.State.Ct)
+}
+
+func TestSetGroupScene(t *testing.T) {
+	b := huego.New(os.Getenv("HUE_HOSTNAME"), os.Getenv("HUE_USERNAME"))
+	id := 1
+	group, err := b.GetGroup(id)
+	if err != nil {
+		t.Fatal(err)
+	}
+	err = group.Scene("2hgE1nGaITvy9VQ")
+	if err != nil {
+		t.Fatal(err)
+	}
+	t.Logf("Scene of group %d set to %s", group.Id, group.State.Scene)
+}
+
+func TestSetGroupTransitionTime(t *testing.T) {
+	b := huego.New(os.Getenv("HUE_HOSTNAME"), os.Getenv("HUE_USERNAME"))
+	id := 1
+	group, err := b.GetGroup(id)
+	if err != nil {
+		t.Fatal(err)
+	}
+	err = group.TransitionTime(10)
+	if err != nil {	
+		t.Fatal(err)
+	}
+	t.Logf("TransitionTime of group %d set to %d", group.Id, group.State.TransitionTime)
+}
+
+func TestSetGroupEffect(t *testing.T) {
+	b := huego.New(os.Getenv("HUE_HOSTNAME"), os.Getenv("HUE_USERNAME"))
+	id := 1
+	group, err := b.GetGroup(id)
+	if err != nil {
+		t.Fatal(err)
+	}
+	err = group.Effect("colorloop")
+	if err != nil {
+		t.Fatal(err)
+	}
+	t.Logf("Effect of group %d set to %s", group.Id, group.State.Effect)
+}
+
+func TestSetGroupAlert(t *testing.T) {
+	b := huego.New(os.Getenv("HUE_HOSTNAME"), os.Getenv("HUE_USERNAME"))
+	id := 1
+	group, err := b.GetGroup(id)
+	if err != nil {
+		t.Fatal(err)
+	}
+	err = group.Alert("lselect")
+	if err != nil {
+		t.Fatal(err)
+	}
+	t.Logf("Alert of group %d set to %s", group.Id, group.State.Alert)
 }