+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Changelog
 
+## 1.0.0-alpha.3
+* Added `Group` receivers: `Alert()`, `Bri()`, `Ct()`, `Effect()`, `Hue()`, `IsOn()`, `Off()`, `On()`, `Rename()`, `Sat()`, `Scene()`, `TransitionTime()` and `Xy()`
+* Added `Light` receivers: `Alert()`, `Effect()` and `TransitionTime()`
+* Renamed type `Action` to `State` that belongs to `Group`
+
 ## 1.0.0
 Initial release